+++ microservice/microservice/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Beam scheduler jobs have a new `SuspendedAt` property
+
+### Changed
+- `BeamScheduler.GetJobs` is obsolete and `GetAllJobs` should be used instead
+
 ## [4.3.0] - 2025-05-08
 
 ### Added